DuckDuckGo plugin for site searching
Adds a search form to the wiki, powered by DuckDuckGo. The quality of
results will depend on if/when DuckDuckGo have indexed your site, and
how recently.

DuckDuckGo market themselves as privacy-concious and claim to not track
your searches or purchases.

This is an alternative to the Google plugin (and is derived from it).