pass PAGE variable to templates
I am not sure why base templates do not know which page they are in,
but this seems like a crucially important information.

I have needed this info to print the path to the current page,
specifically to insert a tracking 1x1 pixel with the path to the
current page for (privacy-aware) analytics with GoatCounter. I have
tried *many* different things, including the trail and parentlinks
plugin, along with the BASEURL and other variables, to no avail. I
couldn't find a single existing variable that would include
the (unique) path to the current page.

I will also note that this variable is refered to in
ikiwiki/directive/template, as PAGE (and also BASENAME), but it seems
those are only available when using the template directive, not in the
base templates used by themes, which also seems silly.