Switch from pubdate=pubdate to class=dt-published
At some point in the past, the "pubdate" attribute was defined for
HTML(5). It is no longer, so generating this attribute results in
invalid HTML.

The intent was to signal that time described corresponded to the
document's publication date. The "h-entry" microformat¹ describes
another way to achieve this: apply the class "dt-published" to the
element. The microformat suggests that this *should* be applied to
a <time> tag but does not forbid other elements. When not in HTML5
mode, we generate a <span> tag. Apply the class in that case too.

¹ http://microformats.org/wiki/h-entry