aggregate enclosures
aggregate: When a feed has an enclosure that is an image, audio, or video,
include the enclosure in the generated page.

The enclosure is hotlinked from the original feed, not copied.

My use case is to include a mastodon rss feed in amoung other rss feeds for
users who don't use mastodon. It could also be used to aggregate together
podcasts, etc.

Other enclosure types than image, audio, video, could be added, perhaps
a generic one? But these are the main ones.

The template uses 50% width for image and video, because often
attachments are in a high resolution, which will default to being
perhaps too wide for the page, or taking up a lot of vertical space. By
making it take up at most half the page width, that is avoided, while
also leaving room for any sidebar.

Sponsored-by: Shae Erisson on Patreon