support non-CGI links in recentchanges
This reverts parts of commit 340fe9707c61064774658e6ec2aec3b07b33c120.

When that was introduced, it was to solve various issues with linking
to pages modified in recent changes. The commit message above, in
particular, says that it can allow new pages to be created if
missing.

According to the discussion in plugins/recentchanges/discussion.md,
however, it seems like it also fixes case folding issues, for which we
do not have a fix here.

But at least this will restore the RSS feed functionality for recent
changes when CGI is disabled. Without this patch, the links do not
point to the pages at all and instead to "anchors" inside the recent
changes page, which makes them much less useful.

In particular, I'm using the recent changes RSS feed to run various
hooks on the blog: a link checker, an archiver script and so on. This
was broken by the switch to CGI, and I have only noticed now because
only now have I disabled CGI on my wiki.